requirements.txt
PyQt5==5.15.10
PyQtChart==5.15.6
numpy==1.26.4
pandas==2.2.2
yfinance==0.2.38
google-generativeai==0.3.2
python-dotenv==1.0.0
pytest==8.0.0
